requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-configurations==2.3.2
django-cors-headers==3.13.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
yapf==0.32.0
